Rename InductionJourneyContainer to InductionStepsContainer.
