+++ packages/webapp/src/inductions/components/induction-journeys/common/index.ts
@@ -1,7 +1,7 @@
 export * from "./endorsements-status";
 export * from "./induction-expires-in";
-export * from "./induction-journey-container";
-export * from "./induction-journey-step";
+export * from "./induction-step";
+export * from "./induction-steps-container";
 export * from "./member-card-preview";
 export * from "./waiting-for-profile";
 export * from "./waiting-for-video";
