+++ packages/webapp/src/inductions/components/induction-journeys/common/induction-step.tsx
